Added spaceship following mouse cursor.

+++ include/input/mouse.hpp
@@ -2,20 +2,20 @@
  * @author Nikita Mochalov (github.com/tralf-strues)
  * @file mouse.hpp
  * @date 2022-05-23
- * 
+ *
  * The MIT License (MIT)
  * Copyright (c) 2022 Nikita Mochalov
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation
  * the rights to use, copy, modify, merge, publish, distribute, sublicense,
  * and/or sell copies of the Software, and to permit persons to whom the
  * Software is furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
  * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -27,15 +27,17 @@
 
 #pragma once
 
+#include <stdint.h>
+
 namespace gwars {
 
 struct MouseMoveEvent
 {
-    int32_t x{0};
-    int32_t y{0};
+    float x{0};
+    float y{0};
 
     MouseMoveEvent() = default;
-    MouseMoveEvent(int32_t x, int32_t y) : x(x), y(y) {}
+    MouseMoveEvent(float x, float y) : x(x), y(y) {}
 };
 
 } // namespace gwars
